eventWithMultipleTagsIsReportedAsPartOfException move to TestSuite
